Move mock to mocks.
