+++ tests/unit/test_crawler_base.py
@@ -4,6 +4,8 @@ import pytest
 
 from databricks.labs.ucx.tacl._internal import CrawlerBase
 
+from .mocks import MockBackend
+
 
 @dataclass
 class Foo:
@@ -11,18 +13,19 @@ class Foo:
     second: bool
 
 
-def test_invalid(mock_backend):
+def test_invalid():
     with pytest.raises(ValueError):
-        CrawlerBase(mock_backend, "a.a.a", "b", "c")
+        CrawlerBase(MockBackend(), "a.a.a", "b", "c")
 
 
-def test_full_name(mock_backend):
-    cb = CrawlerBase(mock_backend, "a", "b", "c")
+def test_full_name():
+    cb = CrawlerBase(MockBackend(), "a", "b", "c")
     assert "a.b.c" == cb._full_name
 
 
-def test_snapshot_appends_to_existing_table(mock_backend):
-    cb = CrawlerBase(mock_backend, "a", "b", "c")
+def test_snapshot_appends_to_existing_table():
+    b = MockBackend()
+    cb = CrawlerBase(b, "a", "b", "c")
     runs = []
 
     def fetcher():
@@ -35,13 +38,13 @@ def test_snapshot_appends_to_existing_table(mock_backend):
     cb._snapshot(Foo, fetcher=fetcher, loader=lambda: [Foo(first="first", second=True)])
 
     insert = "INSERT INTO a.b.c (first, second) VALUES ('first', TRUE)"
-    assert len(mock_backend.queries) == 1
-    assert insert == mock_backend.queries[0]
+    assert len(b.queries) == 1
+    assert insert == b.queries[0]
 
 
-def test_snapshot_appends_to_new_table(mock_backend):
-    mock_backend._fails_on_first = {"INSERT INTO a.b.c": "TABLE_OR_VIEW_NOT_FOUND ..."}
-    cb = CrawlerBase(mock_backend, "a", "b", "c")
+def test_snapshot_appends_to_new_table():
+    b = MockBackend(fails_on_first={"INSERT INTO a.b.c": "TABLE_OR_VIEW_NOT_FOUND ..."})
+    cb = CrawlerBase(b, "a", "b", "c")
     runs = []
 
     def fetcher():
@@ -55,7 +58,7 @@ def test_snapshot_appends_to_new_table(mock_backend):
 
     insert = "INSERT INTO a.b.c (first, second) VALUES ('first', TRUE)"
     create = "CREATE TABLE a.b.c (first STRING, second BOOLEAN) USING DELTA"
-    assert len(mock_backend.queries) == 3
-    assert insert == mock_backend.queries[0]
-    assert create == mock_backend.queries[1]
-    assert insert == mock_backend.queries[2]
+    assert len(b.queries) == 3
+    assert insert == b.queries[0]
+    assert create == b.queries[1]
+    assert insert == b.queries[2]
